requirements_dev.txt
# Tools
black
bump2version
isort
pre-commit
pylint
pytest
tox

# Dependencies
pyserial
